A vote for community Logo: DMFF
It has been nearly half a year since the official release of DMFF. As a Jax-based python package, DMFF provides a fully differentiable implementation of molecular force field models. We have designed a logo for DMFF. Here is the latest version of it. Please determine whether you agree to use it. 

![DMFF logo](https://dp-public.oss-cn-beijing.aliyuncs.com/community/logo/DMFF.jpg)