A vote for community Logo: dflow
It has been nearly half a year since the official release of dflow. As a framework for constructing scientific workflows, dflow provides user-friendly functional programming interfaces and fully supports the fulfillment of complex process logic requirements faced by scientific computing. We have designed a logo for dflow. Here is the latest version of it. Please determine whether you agree to use it. 

![dflow logo](https://dp-public.oss-cn-beijing.aliyuncs.com/community/logo/dflow.jpg)