A vote for community Logo: DeePMD-kit
DeePMD-kit is designed to minimize the effort required to build a deep learning-based model of interatomic potential energy and force field and to perform molecular dynamics (MD). Version 2.1.5 was released in 2022, adding the attention-based DPA-1 model. We have designed a logo for DeePMD-kit. Here is the latest version of it. Please determine whether you agree to use it. 

![DeePMD-kit logo](https://dp-public.oss-cn-beijing.aliyuncs.com/community/logo/DeePMD-kit.png)