A vote for community Logo: DeepFlame
Since the official release of Version 1.0, DeepFlame has made many attempts and achieved milestones in improving the basic functions of combustion and flow simulation, enhancing stability and usability, and realizing high-performance heterogeneous computing. We have designed a logo for DeepFlame. Here is the latest version of it. Please determine whether you agree to use it. 

![DeepFlame Logo](https://dp-public.oss-cn-beijing.aliyuncs.com/community/logo/DeepFlame.png)